+++ smerp/src/main/java/com/domino/smerp/itemorder/dto/request/ItemOrderRequest.java
@@ -2,11 +2,13 @@ package com.domino.smerp.itemorder.dto.request;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
 public class ItemOrderRequest {
     private Long itemId; // 품목 ID
-    private int qty;     // 수량
+    private BigDecimal qty;     // 수량
 }
