+++ smerp/src/main/java/com/domino/smerp/itemorder/dto/response/ItemOrderResponse.java
@@ -4,12 +4,14 @@ import com.domino.smerp.itemorder.ItemOrderCrossedTable;
 import lombok.Builder;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Builder
 public class ItemOrderResponse {
     private Long itemId;
     private String itemName;
-    private int qty;
+    private BigDecimal qty;
 
     public static ItemOrderResponse from(ItemOrderCrossedTable entity) {
         return ItemOrderResponse.builder()
